Move code for cart buttons in a separate component
