+++ src/components/Header.jsx
@@ -1,47 +1,18 @@
 import { useContext, useRef } from 'react';
 import Modal from './common/Modal';
 import Cart from './Cart';
-import { priceFormatter } from '../utils/price-formatter';
 import { CartContext } from '../store/shopping-cart-context';
+import CartButtons from './CartButtons';
 
 function Header() {
   const modalRef = useRef();
-  const { items, total, handleDeleteCart } = useContext(CartContext);
+  const { items } = useContext(CartContext);
 
   function onCartClick() {
     modalRef.current.open();
   }
 
-  function onClearCart(event) {
-    event.preventDefault();
-    handleDeleteCart();
-  }
-
   const cartQuantity = items ? items.length : 0;
-
-  const cartButtons =
-    cartQuantity > 0 ? (
-      <div
-        className="flex gap-x-2 items-center"
-        data-testid="cart-buttons-container"
-      >
-        <p className="font-quicksand font-bold" data-testid="cart-total">
-          Your Total:{' '}
-          <span className="text-red-700">{priceFormatter.format(total)}</span>
-        </p>
-        <div className="flex gap-x-2">
-          <button onClick={onClearCart} data-testid="clear-cart-btn">
-            Clear Cart
-          </button>
-          <button data-testid="close-cart-btn">Close</button>
-          <button onClick={() => {}} data-testid="checkout-btn">
-            Checkout
-          </button>
-        </div>
-      </div>
-    ) : (
-      <button data-testid="close-cart-btn">Close</button>
-    );
 
   return (
     <>
@@ -49,7 +20,7 @@ function Header() {
         ref={modalRef}
         title="Cart"
         modalContent={<Cart />}
-        modalButtons={cartButtons}
+        modalButtons={<CartButtons />}
       />
       <header className="flex justify-between items-center px-[15%] py-10 bg-gradient-to-r from-[#31220b] to-[#382e1b]">
         <div className="flex gap-x-10 items-center">
